component, probe, timer modules

+++ src/adc.h
@@ -1,4 +1,6 @@
 #pragma once
 
+#include <stdint.h>
+
 void adc_init(void);
-unsigned int adc_average(unsigned int channel, unsigned int num);
+uint_fast16_t adc_average(uint_fast8_t channel, uint_fast16_t num);
